# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Corridor": {
        "additionalProperties": false,
        "description": "A road, as the parameters needed to fetch and segment it again.\n\nDeliberately *not* the resolved geometry. Geometry belongs to a run, because the\nOSM extract behind it changes: two runs of the same corridor a month apart are two\ndifferent centrelines and must not be conflated. What is stable is the request \u2014\nthis reference, this bounding box, this unit length.",
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "unit_length_m": {
            "default": 500.0,
            "title": "Unit Length M",
            "type": "number"
          }
        },
        "required": [
          "tenant_id",
          "project_id",
          "name"
        ],
        "title": "Corridor",
        "type": "object"
      },
      "CorridorCreate": {
        "additionalProperties": false,
        "description": "Created under the project in the path. `project_id` is never in the body.",
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "description": "south, west, north, east, in degrees.",
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "ref": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Road reference as OSM knows it, e.g. 'B9'. Null for a client-supplied centreline.",
            "title": "Ref"
          },
          "unit_length_m": {
            "default": 500.0,
            "exclusiveMinimum": 0.0,
            "maximum": 100000.0,
            "title": "Unit Length M",
            "type": "number"
          }
        },
        "required": [
          "name"
        ],
        "title": "CorridorCreate",
        "type": "object"
      },
      "CorridorPatch": {
        "additionalProperties": false,
        "description": "See :class:`ProjectPatch`. `project_id` is not editable \u2014 see the store.",
        "properties": {
          "bbox": {
            "anyOf": [
              {
                "description": "south, west, north, east, in degrees.",
                "maxItems": 4,
                "minItems": 4,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "ref": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "unit_length_m": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Length M"
          }
        },
        "title": "CorridorPatch",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The response body of every refusal this API makes.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Refusal"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "description": "What kind of refusal this is, as something a client can branch on.\n\nCodes rather than status alone, because 422 is doing two jobs \u2014 a malformed request\nbody and a panel that breaks the input contract are both 422 and are not the same\nproblem. A client retries the first after fixing its JSON and the second after\nfixing its data.",
        "enum": [
          "tenant_required",
          "invalid_request",
          "contract_violation",
          "payload_rejected",
          "not_found",
          "in_use",
          "artefact_unavailable",
          "too_large",
          "engine_refused",
          "internal"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "description": "One thing wrong with the request, and where.",
        "properties": {
          "location": {
            "description": "Dotted path into the request, e.g. 'body.name'.",
            "title": "Location",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Refusal": {
        "additionalProperties": false,
        "description": "Why the request was refused, in the terms the caller can act on.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "errors": {
            "description": "Every fault found, when a request had more than one.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The single thing at fault, when there is one.",
            "title": "Field"
          },
          "message": {
            "description": "Written for a person. Names the column, the id or the setting.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Refusal",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nCorridor road-risk assessment from open data, with provenance.\n\n**Two things this deployment does not yet do, stated here rather than discovered.**\n\n* **Nothing executes jobs.** `POST /jobs` stores a job and returns `202`. It stays\n  `queued` until step 5.1d attaches a runner. `GET /health` reports `runner: null`.\n* **`X-Tenant-Id` is not authentication.** It scopes every read to one tenant, which is\n  what keeps two clients' runs apart, but nothing verifies the claim. Step 5.4a\n  replaces it with real identities and row-level policies in the database.\n\n**A refusal is a result, not an error.** A panel that breaks the input contract is a\n`422` naming the column, and no job is created. An assessment that descended to Mode B,\ndropped a term or declined to score an unsourced weight is a **`200`** \u2014 those are\nfindings the run carries, not failures. Infrastructure breaking is a job whose status is\n`failed`, with a cause.\n\nPayload shape version `1.0`. Factors, tiers and licences come from\n`factors.yaml`; `GET /registry` serves it with the hash of the file it was read from.\n",
    "summary": "Corridor risk from open data, with every number traceable.",
    "title": "Road Risk Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/corridors/{corridor_id}": {
      "delete": {
        "description": "Refused while a job or a run points at it, even though the schema would allow it.\n\n`job.corridor_id` and `run.corridor_id` are `ON DELETE SET NULL`, so nothing would\nbe destroyed here \u2014 but the *link* would be, silently. A run keeps its geometry\ninside the payload and would lose the road it was filed against, which is exactly\nthe kind of quiet loss this project spends its effort preventing.",
        "operationId": "delete_corridor_corridors__corridor_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "corridor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Corridor Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "409": {
            "description": "A job or a run still points at it."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Delete a corridor nothing references",
        "tags": [
          "corridors"
        ]
      },
      "get": {
        "operationId": "get_corridor_corridors__corridor_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "corridor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Corridor Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Corridor"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "One corridor",
        "tags": [
          "corridors"
        ]
      },
      "patch": {
        "operationId": "patch_corridor_corridors__corridor_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "corridor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Corridor Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorridorPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Corridor"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Change the reference, the box or the unit length",
        "tags": [
          "corridors"
        ]
      }
    },
...
    "/projects/{project_id}/corridors": {
      "get": {
        "operationId": "list_corridors_projects__project_id__corridors_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Corridor"
                  },
                  "title": "Response List Corridors Projects  Project Id  Corridors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Every corridor in a project",
        "tags": [
          "corridors"
        ]
      },
      "post": {
        "description": "Create it, or refuse if the project is not this tenant's.\n\nThe store checks the parent against `(tenant_id, project_id)` together, so a\ncorridor cannot be filed under another tenant's project even with a valid id for\nit \u2014 that is a row Postgres refuses, not a rule this route has to remember.",
        "operationId": "create_corridor_projects__project_id__corridors_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorridorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Corridor"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Declare a road under a project",
        "tags": [
          "corridors"
        ]
      }
    },
...
  }
}